+++ Frontend/app/components/molecules/Navbar.tsx
