+++ Frontend/app/components/molecules/StatDisplay.tsx
